+++ src/app/map/page.tsx
@@ -1,14 +1,19 @@
 import type { Metadata } from 'next';
 
+import Sidebar from '@/containers/sidebar';
+
+import Map from '@/doc-containers/map-implementation';
 export const metadata: Metadata = {
   title: 'Map - Open-Earth-Monitor Cyberinfrastructure',
   description: '...',
 };
 
-const Map: React.FC = () => (
-  <div>
-    <h1>Map</h1>
-  </div>
+const MapPage: React.FC = () => (
+  <main className="relative flex h-screen w-screen flex-1">
+    <Sidebar />
+
+    <Map />
+  </main>
 );
 
-export default Map;
+export default MapPage;