+++ src/components/map/legend/index.tsx
@@ -1,9 +1,9 @@
 import { FC, useCallback, useState, PropsWithChildren, useId } from 'react';
 
 import cx from 'clsx';
+import LEGEND_SVG from 'svgs/map/legend.svg?sprite';
 
 import Icon from 'components/icon';
-import LEGEND_SVG from 'svgs/map/legend.svg?sprite';
 import ARROW_DOWN_SVG from 'svgs/ui/arrow-down.svg?sprite';
 
 import SortableList from './sortable/list';